refactor(api): Add subpackages to controller packages
