+++ backend/FleetForge/src/main/java/com/team18/FleetForge/dto/ride/lifecycle/RideCancellationRequestDTO.java
@@ -1,8 +1,6 @@
 package com.team18.FleetForge.dto.ride.lifecycle;
 
-import com.team18.FleetForge.model.enums.RideCancellationRole;
 import lombok.Data;
-import java.time.LocalDateTime;
 
 @Data
 public class RideCancellationRequestDTO {
